[modify]fishData.jsとUIManager.jsを作成し、FishingScene.jsの処理を分担させた。
